Broken: Wrong params
I am beginning to think Scrapy might be the wrong choice. While it's true that
we need to scrape a lot of stuff, we need a different spider for each page, and
we only need to collect a few values.

It might make sense to just use something simpler to automate form filing out.

Then again, maybe I just need to work with scrapy for a bit longer?

Yeah, I should get a working example first, then go from there.